Move BulkUpdateFlow to common
