+++ common/src/main/scala/weco/concepts/common/elasticsearch/BulkUpdateFlow.scala
@@ -1,10 +1,9 @@
-package weco.concepts.aggregator
+package weco.concepts.common.elasticsearch
 
 import akka.NotUsed
 import akka.stream.scaladsl.Flow
 import grizzled.slf4j.Logging
 import org.elasticsearch.client.{Response, ResponseException}
-import weco.concepts.common.elasticsearch.Indexer
 
 import scala.util.{Failure, Success}
